+++ plugin/android-junit5/src/main/templates/Libraries.kt
@@ -6,4 +6,6 @@ internal object Libraries {
     const val instrumentationCore = "@INSTRUMENTATION_GROUP@:@INSTRUMENTATION_CORE@"
     const val instrumentationExtensions = "@INSTRUMENTATION_GROUP@:@INSTRUMENTATION_EXTENSIONS@"
     const val instrumentationRunner = "@INSTRUMENTATION_GROUP@:@INSTRUMENTATION_RUNNER@"
+
+    const val junitPlatformLauncher = "@JUNIT_PLATFORM_LAUNCHER@"
 }